New code

+++ src/app/login/login.component.ts
@@ -45,27 +45,36 @@ export class LoginComponent implements OnInit {
     this.isProcess = true;
     this.database.login(this.submitForm.value.nickname).then((res) =>{
       this.isProcess = false;
+      // console.log(res.val())
         if (res.val() == null){
           this.openDialog("Wrong Nickname or Password");
           return;
-        } else if (res.val().length > 1){
-          this.openDialog("Error");
-          return;
         }
         else{
-          let user =  <User>res.val()[0];
-          console.log(user)
-          if (this.checkPassword(user, this.submitForm.value.password)){
-            if (this.cookieService.check("sang-app-chat")){
-              this.cookieService.delete("sang-app-chat");
+          let user = null;
+          res.val().forEach(element => {
+            if (element){
+             user =  <User>element;
             }
-              var ob = {user_id: user.user_id}
-              this.cookieService.set("sang-app-chat", JSON.stringify(ob), 4, '/');
-            this.router.navigate(['/chat-page']);
-          }
-          else
-          this.openDialog("Wrong Nickname or Password");
-          return;
+          });
+          
+          if (user != null){
+            if (this.checkPassword(user, this.submitForm.value.password)){
+              if (this.cookieService.check("sang-app-chat")){
+                this.cookieService.delete("sang-app-chat");
+              }
+                var ob = {user_id: user.user_id}
+                this.cookieService.set("sang-app-chat", JSON.stringify(ob), 4, '/');
+              this.router.navigate(['/chat-page']);
+            }
+            else
+            this.openDialog("Wrong Nickname or Password");
+            return;
+          } else{
+            this.openDialog("Error");
+          }  
+
+
         }
     })
